backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-cors-headers>=3.10.0
djangorestframework-simplejwt
django-filter
psycopg2
dj-database-url
python-dotenv
Pillow
gunicorn
boto3 
django-storages
faker
channels[daphne]





